Work around RSA SHA-2 signature issues in conch
This was supposed to be fixed in Twisted upstream
(https://twistedmatrix.com/trac/ticket/9765), and that fix is in Debian
now.  However, regression tests still seem to fail in GitLab CI but not
locally (see e.g.
https://salsa.debian.org/ssh-team/openssh/-/jobs/3513178).  Leave this
in place for now until we figure out what's wrong.

Forwarded: not-needed
Last-Update: 2022-11-14

Patch-Name: conch-ssh-rsa.patch