Fix performance of regression tests with UsePAM
When running the regression tests with `UsePAM yes` as we now do,
performance is terrible due to DNS resolution delays introduced by ssh
setting `PAM_RHOST` to `UNKNOWN`.  The only even semi-reasonable way
I've found to work around this is to make `UNKNOWN` resolvable while
running the regression tests, so do that.