Install tmpfiles.d to avoid cleanup of ssh-agent socket in /tmp/
sd-tmpfiles will clean up old files/dirs from /tmp/ on a timer,
which might remove the ssh-agent socket if it's been running for
a long time.

Ideally ssh-agent should switch to using XDG_RUNTIME_DIR, or socket
based per-session activation as per #1068416 or at least take a flock
on the socket while it's running as per tmpfiles.d protocol. For now
as a quick workaround ship a tmpfiles.d snippet that adds an exception
for /tmp/ssh-* to avoid issues. This should be reverted once ssh-agent
changes to employ any of the other mentioned solutions.

Closes: #1068416