Drop most of Salsa CI's suffix from version string
The "integrity" regression tests are somewhat sensitive to the length of
the server banner (which is sent before key exchange), and the long
suffixes used for CI builds throw them off too much.  Since the final
segment of Salsa CI version numbers is a sequence number that runs
independently in different forks, this can mean that the exact same
commit fails in forks despite passing in `ssh-team/openssh`, which is
extremely confusing.

Having some indication that a build comes from CI is somewhat useful,
but we don't need to embed all the fine details: a plain `+salsaci`
suffix is good enough for that.