Add sshd-keygen service
Runs only on first boot (empty /etc/) and if /etc/ exists,
is writable and is not a symlink. Allows a usr-only image
with an empty /etc/ (firstboot or factory reset) to boot and
have a working sshd out of the box.
Given the firstboot condition is checked, it will be a no-op and
gracefully skipped on normal systems where /etc/ exists and
is created by the installer.